requirements.txt
datasets==3.6.0
pyyaml>=6.0
mlflow==2.22.4
numpy==2.1.1
pandas==2.2.3
scikit-learn==1.8.0
matplotlib>=3.5
xgboost==3.1.3

ipykernel==6.25.2
ipython==8.12.2
notebook==6.4.10
traitlets==5.9.0
setuptools==80.10.1
